requirements.txt
h5py==3.1.0
ipython==6.4.0
numpy==1.19.5
Pillow==7.0.0
PyWavelets==1.1.1
scikit-image==0.17.2
scikit-learn==0.22.1
scipy==1.2.0

torch==1.7.1+cu110
torchvision==0.8.2+cu110